Purchase creation done

+++ src/IMS/core/IMS.BusinessModel/ViewModel/PurchaseDetailsModel.cs
@@ -4,8 +4,8 @@
     {  
         public bool IsDeleted { get; set; }
         public long ProductId { get; set; }
-        public decimal Quantity { get; set; }
-        public int UnitPrice { get; set; }
+        public int Quantity { get; set; }
+        public decimal UnitPrice { get; set; }
         public decimal Total { get; set; }
     }
 }
